Fixed occasional crash related to Goron spawning
Null pointer caused by block not found, probably if Goron's position is
in a non-existant chunk. Fixed by using try/catch to simply ignore error
and move on, since it's not critical for each Goron to spawn and the
error seemed to only happen twice per world load on super flat setting.