Fix possible crash rendering blocks without items
Apparently it is possible for ItemDungeonStone to contain an incorrect
render state, i.e. for a block such a Stone Slab that does not have a
corresponding Item model.

Should only affect games started on earlier versions and had someone make
a Dungeon Stone item render as one of these invalid blocks, then kept the
item version in their inventory (blocks in the world render using the
block model, so are unaffected by this bug).