Fixes issue with PlayerUtils#consumeHeldItem
- Consuming held item with this method caused helm or other slot to
duplicate the new ItemStack on the client, e.g. if the item was consumed
entirely the helm would seem to disappear. Clicking on the slot would
re-sync and correctly show the helm. Seems to be an issue with the slot
packet.

- Furthermore, sending the slot packet in #consumeInventoryItem, while
it doesn't break anything like the above, it also doesn't seem to have
any effect, i.e. the issue from commit 81e3991 seems to be a non-issue.