Fix infinite Wizzrobe-teleportation loop
Since there is not yet any EnderTeleportEvent.Post, any mod that causes
damage during teleportation does so when the teleportation may not
actually be successful.

Since Wizzrobes teleport when attacked, receiving damage during
teleportation caused an infinite loop.

This will no longer happen, but the tele-event is still posted 64 times
(for Endermen as well) - each time it is canceled, but there is not
currently any way to break out of the loop.