Fix weird config category comment issue
Apparently adding a category comment after one or more entries has been added for that category causes weird things to happen - the category comment would alternate between existing and not, and the map making category would get duplicated.

In 1.8, the case of the category name also caused this issue and had to be
made lower case. Go figure.