Fix potential memory issue when a World unloads
The ZSS structure generator instances are stored in the event handler class, each of which contains a reference to the last World object to use it.

Should that World be unloaded but the game is still going (for example the Nether), the reference stored in the still-existing ZSS generator instance would prevent the World object from being garbage collected.