WIP randomized rupee trades
The plan is for these to be used to add new trades to merchant's inventory as time goes on, similar to vanilla traders.

Could be cool to have them usable for the initial lists as well, e.g. select min to max (e.g. 2 to 4) randomized entries from the weighted list when populating for the first time.